Set data-app in custom Vuetify mount command
Default Vuetify behavior in certain components (like v-menu, v-autocomplete) is to attach a popup to an ancestor with a `data-app` attribute, the same element that has the `v-application` class. I added the suggested change to my command and am able to successfully test a component that uses a v-autocomplete, so I thought it might be handy in this repo.